Cint quiz app
This commit represents a React application created with Vite, using Tailwind/HeadlessUI to provide a quiz to the user.

React router is used for routing and react query for data control.

A context  and reducer are used for injecting state between pages.

A small number of tests are including but by no means comprehensive.